+++ src/Git_Updater/Additions/Additions.php
@@ -48,6 +48,7 @@ class Additions {
 		}
 
 		$this->add_headers( $config, $repos, $type );
+		$this->add_source( $config );
 
 		return true;
 	}
@@ -138,16 +139,40 @@ class Additions {
 	 * @return array
 	 */
 	public function deduplicate( $options ) {
-		$list_plugin_addons = $this->get_repo_cache( 'git_updater_repository_add_plugin' );
-		$list_plugin_addons = ! empty( $list_plugin_addons['git_updater_repository_add_plugin'] ) ? $list_plugin_addons['git_updater_repository_add_plugin'] : [];
+		if ( empty( $options ) ) {
+			return $options;
+		}
 
-		$list_theme_addons = $this->get_repo_cache( 'git_updater_repository_add_theme' );
-		$list_theme_addons = ! empty( $list_theme_addons['git_updater_repository_add_theme'] ) ? $list_theme_addons['git_updater_repository_add_theme'] : [];
+		$plugin_addons = $this->get_repo_cache( 'git_updater_repository_add_plugin' );
+		$plugin_addons = ! empty( $plugin_addons['git_updater_repository_add_plugin'] ) ? $plugin_addons['git_updater_repository_add_plugin'] : [];
 
-		$options = array_merge( $options, $list_plugin_addons, $list_theme_addons );
+		$theme_addons = $this->get_repo_cache( 'git_updater_repository_add_theme' );
+		$theme_addons = ! empty( $theme_addons['git_updater_repository_add_theme'] ) ? $theme_addons['git_updater_repository_add_theme'] : [];
+
+		$packages = array_merge( $plugin_addons, $theme_addons );
+
+		foreach ( $packages as $key => $item ) {
+			foreach ( $options as $option ) {
+				if ( $item['ID'] === $option['ID'] && $item['source'] !== $option['source'] ) {
+					unset( $packages[ $key ] );
+				}
+			}
+		}
+
+		$collections = get_site_option( 'git_updater_collections' );
+		$collections = $collections ?: [];
+		foreach ( $collections as $collection ) {
+			foreach ( $options as $key => $item ) {
+				if ( $item['source'] === $collection['ID'] ) {
+					unset( $options[ $key ] );
+				}
+			}
+		}
+
+		$options = array_merge( $options, $packages );
 		foreach ( array_keys( $options ) as $key ) {
-			unset( $options[ $key ]['source'] );
-			$options[ $key ]['release_asset'] = ! empty( $options[ $key ]['release_asset'] ) ? true : false;
+			$options[ $key ]['release_asset']   = ! empty( $options[ $key ]['release_asset'] ) ? true : false;
+			$options[ $key ]['private_package'] = ! empty( $options[ $key ]['private_package'] ) ? true : false;
 			ksort( $options[ $key ] );
 		}
 		$options = array_map( 'unserialize', array_unique( array_map( 'serialize', $options ) ) );