+++ src/Git_Updater/Additions/Settings.php
@@ -91,15 +91,14 @@ class Settings {
 			'git_updater_additions' === $post_data['option_page']
 		) {
 			$new_options = $post_data['git_updater_additions'] ?? [];
-
 			$new_options = $this->sanitize( $new_options );
+			$bad_input   = empty( $new_options[0]['slug'] ) || empty( $new_options[0]['uri'] );
 
 			foreach ( $options as $option ) {
 				$is_plugin_slug = preg_match( '@/@', $new_options[0]['slug'] );
-				$type_plugin    = \preg_match( '/plugin/', $new_options[0]['type'] );
+				$type_plugin    = preg_match( '/plugin/', $new_options[0]['type'] );
 				$bad_input      = $type_plugin && ! $is_plugin_slug;
 				$bad_input      = ! $bad_input ? ! $type_plugin && $is_plugin_slug : $bad_input;
-				$bad_input      = $bad_input || empty( $new_options[0]['slug'] ) || empty( $new_options[0]['uri'] );
 				$duplicate      = in_array( $new_options[0]['ID'], $option, true );
 				if ( $duplicate || $bad_input ) {
 					$_POST['action'] = false;
@@ -182,7 +181,7 @@ class Settings {
 
 		add_settings_section(
 			'git_updater_additions',
-			esc_html__( 'Additions', 'git-updater' ),
+			esc_html__( 'Addition Packages', 'git-updater' ),
 			[ $this, 'print_section_additions' ],
 			'git_updater_additions'
 		);
@@ -252,6 +251,19 @@ class Settings {
 				'title'   => __( 'Check if a release asset is required.', 'git-updater' ),
 			]
 		);
+
+		add_settings_field(
+			'private_package',
+			esc_html__( 'Private Package', 'git-updater' ),
+			[ $this, 'callback_checkbox' ],
+			'git_updater_additions',
+			'git_updater_additions',
+			[
+				'id'      => 'git_updater_additions_private_package',
+				'setting' => 'private_package',
+				'title'   => __( 'Check if this package is not to be shared with aggregators.', 'git-updater' ),
+			]
+		);
 	}
 
 	/**
@@ -267,8 +279,10 @@ class Settings {
 		foreach ( (array) $input as $key => $value ) {
 			$new_input[0][ $key ] = 'uri' === $key ? untrailingslashit( esc_url_raw( trim( $value ) ) ) : sanitize_text_field( $value );
 		}
-		$new_input[0]['primary_branch'] = ! empty( $new_input[0]['primary_branch'] ) ? $new_input[0]['primary_branch'] : 'master';
-		$new_input[0]['ID']             = md5( $new_input[0]['slug'] );
+		$new_input[0]['primary_branch']  = ! empty( $new_input[0]['primary_branch'] ) ? $new_input[0]['primary_branch'] : 'master';
+		$new_input[0]['ID']              = md5( $new_input[0]['slug'] );
+		$new_input[0]['source']          = md5( home_url() );
+		$new_input[0]['private_package'] = ! empty( $new_input[0]['private_package'] ) ? true : false;
 
 		return $new_input;
 	}