+++ src/Git_Updater/Base.php
@@ -484,6 +484,9 @@ class Base {
 			if ( isset( $hook_extra['plugin'] ) ) {
 				$slug       = dirname( $hook_extra['plugin'] );
 				$new_source = trailingslashit( $remote_source ) . $slug;
+			} elseif ( wp_doing_ajax() && check_ajax_referer( 'updates' ) && isset( $_POST['slug'] ) ) {
+				$slug       = sanitize_key( wp_unslash( $_POST['slug'] ) );
+				$new_source = trailingslashit( $remote_source ) . $slug;
 			}
 		}
 
@@ -494,6 +497,9 @@ class Base {
 			$upgrader_object = Singleton::get_instance( 'Theme', $this );
 			if ( isset( $hook_extra['theme'] ) ) {
 				$slug       = $hook_extra['theme'];
+				$new_source = trailingslashit( $remote_source ) . $slug;
+			} elseif ( wp_doing_ajax() && check_ajax_referer( 'updates' ) && isset( $_POST['slug'] ) ) {
+				$slug       = sanitize_key( wp_unslash( $_POST['slug'] ) );
 				$new_source = trailingslashit( $remote_source ) . $slug;
 			}
 		}